CRUD should properly work now

+++ SomerenApp/Models/Lecturer.cs
@@ -5,11 +5,11 @@
         public int LecturerNumber { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int Age { get; set; }
+        public byte Age { get; set; }
         public string PhoneNumber { get; set; }
         public int RoomId { get; set; }
 
-        public Lecturer(int lecturerNumber, string firstName, string lastName, int age, string phoneNumber, int roomId)
+        public Lecturer(int lecturerNumber, string firstName, string lastName, byte age, string phoneNumber, int roomId)
         {
             LecturerNumber = lecturerNumber;
             FirstName = firstName;
@@ -19,9 +19,6 @@
             RoomId = roomId;
         }
 
-        public Lecturer()
-        {
-            
-        }
+        public Lecturer(){}
     }
 }
